// src/api/generated/schemas/notificationListItemResponse.ts
/**
 * Generated by orval v7.12.2 🍺
 * Do not edit manually.
 * Notifiable API
 * A simple example ASP.NET Core Web API
 * OpenAPI spec version: v1
 */
import type { NotificationChannelType } from './notificationChannelType';
import type { NotificationStatus } from './notificationStatus';
import type { NotificationPriority } from './notificationPriority';

export interface NotificationListItemResponse {
  id?: string;
  channel?: NotificationChannelType;
  status?: NotificationStatus;
  priority?: NotificationPriority;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  body?: string | null;
  createdAt?: string;
  /** @nullable */
  lastUpdatedAt?: string | null;
  /** @nullable */
  correlationId?: string | null;
  isRead?: boolean;
}