// src/api/generated/schemas/notificationReadState.ts
/**
 * Generated by orval v7.12.2 🍺
 * Do not edit manually.
 * Notifiable API
 * A simple example ASP.NET Core Web API
 * OpenAPI spec version: v1
 */

export type NotificationReadState =
  (typeof NotificationReadState)[keyof typeof NotificationReadState];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const NotificationReadState = {
  All: 'all',
  Read: 'read',
  Unread: 'unread',
} as const;